Store full assets individually via assetEmitted tap (#37)
It is possible for assets to be added to the compilation object (e.g.
via compilation.emitAsset) after this plugin's tap of the emit hook is
called. For example, a different plugin can tap the emit hook later on
and call compilation.emitAsset.

By using the assetEmitted hook, we ensure that each new asset is
added to our WeakMap as it is emitted which ensures that all assets
wind up available when we create the assets value in our
Compilation.assets resolver.